refactor(auth): move login/register from SSR page.tsx to client components
